requirements.txt
pip~=25.2
distro~=1.9.0
bcrypt~=4.0.1
fastapi~=0.117.1
SQLAlchemy~=2.0.44
pydantic~=2.11.9
uvicorn~=0.37.0
python-dotenv~=1.1.1
starlette~=0.48.0
openai~=1.108.2
pydub~=0.25.1
itsdangerous